+++ client/coral-framework/graphql/mutations.js
@@ -1,7 +1,6 @@
 import { gql } from 'react-apollo';
 import withMutation from '../hocs/withMutation';
 import update from 'immutability-helper';
-import moment from 'moment';
 
 function convertItemType(item_type) {
   switch (item_type) {
@@ -638,115 +637,6 @@ export const withChangePassword = withMutation(
         return mutate({
           variables: {
             input,
-          },
-        });
-      },
-    }),
-  }
-);
-
-export const withRequestAccountDeletion = withMutation(
-  gql`
-    mutation RequestAccountDeletion {
-      requestAccountDeletion {
-        ...RequestAccountDeletionResponse
-      }
-    }
-  `,
-  {
-    props: ({ mutate }) => ({
-      requestAccountDeletion: () => {
-        return mutate({
-          variables: {},
-          update: proxy => {
-            const RequestAccountDeletionQuery = gql`
-              query Talk_CancelAccountDeletion {
-                me {
-                  id
-                  scheduledDeletionDate
-                }
-              }
-            `;
-
-            const prev = proxy.readQuery({
-              query: RequestAccountDeletionQuery,
-            });
-
-            const scheduledDeletionDate = moment()
-              .add(12, 'hours')
-              .toDate();
-
-            const data = update(prev, {
-              me: {
-                scheduledDeletionDate: { $set: scheduledDeletionDate },
-              },
-            });
-
-            proxy.writeQuery({
-              query: RequestAccountDeletionQuery,
-              data,
-            });
-          },
-        });
-      },
-    }),
-  }
-);
-
-export const withRequestDownloadLink = withMutation(
-  gql`
-    mutation RequestDownloadLink {
-      requestDownloadLink {
-        ...RequestDownloadLinkResponse
-      }
-    }
-  `,
-  {
-    props: ({ mutate }) => ({
-      requestDownloadLink: () => {
-        return mutate({
-          variables: {},
-        });
-      },
-    }),
-  }
-);
-
-export const withCancelAccountDeletion = withMutation(
-  gql`
-    mutation RequestDownloadLink {
-      cancelAccountDeletion {
-        ...CancelAccountDeletionResponse
-      }
-    }
-  `,
-  {
-    props: ({ mutate }) => ({
-      cancelAccountDeletion: () => {
-        return mutate({
-          variables: {},
-          update: proxy => {
-            const CancelAccountDeletionQuery = gql`
-              query Talk_CancelAccountDeletion {
-                me {
-                  id
-                  scheduledDeletionDate
-                }
-              }
-            `;
-
-            const prev = proxy.readQuery({ query: CancelAccountDeletionQuery });
-
-            const data = update(prev, {
-              me: {
-                scheduledDeletionDate: { $set: null },
-              },
-            });
-
-            proxy.writeQuery({
-              query: CancelAccountDeletionQuery,
-              data,
-            });
           },
         });
       },
